refactor: split cli/integrations.ts into directory with per-integration files
